+++ src/main/java/com/job/back/BackApplication.java
@@ -9,5 +9,5 @@ public class BackApplication {
 	public static void main(String[] args) {
 		SpringApplication.run(BackApplication.class, args);
 	}
-
+	
 }
