+++ src/main/java/com/job/back/common/constant/ApiPattern.java
@@ -8,5 +8,6 @@ public interface ApiPattern {
     public static final String AUTH = "/auth";
     public static final String APPLICANT = "/applicant";
     public static final String USER_FILE = "/user-file";
+    public static final String SEARCH = "/search";
     
 }
